0.0014 - Models, Anim, Bug Fix
Added Snap sound effect so there are 6 default sounds. Defaulted one to each tower type.
Changed basic Source and Bouncer visual models to reflect Directionality
Added bouncing animation when tower is triggered
Fixed multiplication by int causing beat to stutter / skip every once in while

+++ Assets/Scripts/Music/TempoHandler.cs
@@ -42,8 +42,8 @@ public class TempoHandler : MonoBehaviour
         //sampleRate = AudioSettings.outputSampleRate;
         //nextTick = startDSPTime * sampleRate;
         //running = true;
-        beatLength = 60.0 / bpm * 4.0f / signatureLo;
-        barLength = beatLength * signatureHi;
+        beatLength = 60.0 / bpm * 4.0 / (double) signatureLo;
+        barLength = beatLength * (double) signatureHi;
         nextBeatTime = startDSPTime + beatLength;
     }
